requirements.txt
flet>=0.9.0
cx_Oracle>=8.3.0
python-dotenv
pytest

